[REF/#623] Refactor MainAppState and move NetworkMonitor to core (#626)
